@@ -13,41 +13,42 @@ interface IMobilePhoneInputField {
   value?: any;
   disabled?: boolean;
 }
-const MobilePhoneInputField: FunctionComponent<IMobilePhoneInputField> =
-  props => {
-    const [field, { error }, helpers] = useField(props);
-    return (
-      <div
-        className={`nhsuk-form-group ${error && "nhsuk-form-group--error"} ${
-          props.hidelabel && "hide-label"
-        }`}
+const MobilePhoneInputField: FunctionComponent<
+  IMobilePhoneInputField
+> = props => {
+  const [field, { error }, helpers] = useField(props);
+  return (
+    <div
+      className={`nhsuk-form-group ${error && "nhsuk-form-group--error"} ${
+        props.hidelabel && "hide-label"
+      }`}
+    >
+      {props.label && (
+        <label htmlFor="mobilePhoneNumber" className="nhsuk-label">
+          {props.label}
+        </label>
+      )}
+      <PhoneInput
+        placeholder="Enter mobile phone number"
+        onBlur={field.onBlur}
+        disabled={props.disabled}
+        name="mobilePhoneNumber"
+        defaultCountry="GB"
+        onChange={value => {
+          helpers.setValue(value);
+        }}
+      />
+      <span
+        className="nhsuk-error-message"
+        id="confirmTOTPCode--error-message"
+        role="alert"
       >
-        {props.label && (
-          <label htmlFor="mobilePhoneNumber" className="nhsuk-label">
-            {props.label}
-          </label>
-        )}
-        <PhoneInput
-          placeholder="Enter mobile phone number"
-          onBlur={field.onBlur}
-          disabled={props.disabled}
-          name="mobilePhoneNumber"
-          defaultCountry="GB"
-          onChange={value => {
-            helpers.setValue(value);
-          }}
-        />
-        <span
-          className="nhsuk-error-message"
-          id="confirmTOTPCode--error-message"
-          role="alert"
-        >
-          <span className="nhsuk-u-visually-hidden">Error: </span>
-          {error}
-        </span>
-        <InputFooterLabel label={props.footer || ""} />
-      </div>
-    );
-  };
+        <span className="nhsuk-u-visually-hidden">Error: </span>
+        {error}
+      </span>
+      <InputFooterLabel label={props.footer || ""} />
+    </div>
+  );
+};
 
 export default connect(MobilePhoneInputField);
