@@ -8,7 +8,7 @@ export const StringValidationSchema = (
     .string()
     .nullable()
     .transform(value => {
-      if (value) return (value as string).trim();
+      if (value && value.length > 0) return (value as string).trim();
     })
     .min(1)
     .max(maxLength, `${fieldName} must be shorter than ${maxLength} characters`)
